Update the security group in Creation of DB
In the Set up Networking step we asked customer to create a security group with the name DB-SG. That Security group has an inbound rule to allow the connectivity with the target database from Replica instance. In this step we are asking the customer to create again the Security Group and giving them the example name as it was the one created in the previous step. Also, we say to the customer that we will modify the security group later. The /db-mig/6_db.png shows a screenshot with "Existing security group" selected and not created. That is the reason of my suggestion of this to be addressed, to have consistency between text, image and previous steps. Thanks